aeboards/ext65 refactor (#10820)
* ext65 refactor

* Update rev2.c

* update keymap

* Update keyboards/aeboards/ext65/readme.md

* update with suggestions
