+++ keyboards/aeboards/ext65/ext65.c
@@ -1,0 +1,1 @@
+#include "ext65.h"
