+++ keyboards/aeboards/ext65/ext65.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "quantum.h"
+
+#if defined(KEYBOARD_aeboards_ext65_rev1)
+    #include "rev1.h"
+#elif defined(KEYBOARD_aeboards_ext65_rev2)
+    #include "rev2.h"
+#endif
