+++ keyboards/aeboards/ext65/readme.md
@@ -1,0 +1,19 @@
+# EXT65
+
+A southpaw inspired keyboard by [AEBoards](https://aeboards.com/)
+
+* Keyboard Maintainer: [Xelus22](https://github.com/Xelus22)
+* Hardware Supported: EXT65 Rev1, Rev2
+* Hardware Availability: Custom keyboard group buys
+
+Make example for this keyboard (after setting up your build environment):
+
+    make aeboards/ext65/rev1:default
+    make aeboards/ext65/rev2:default
+
+Flashing example for this keyboard:
+
+    make aeboards/ext65/rev1:default:flash
+    make aeboards/ext65/rev2:default:flash
+
+See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
