--- keyboards/aeboards/ext65/rev1/keymaps/default/readme.md
@@ -1,2 +1,0 @@
-# The Default Ext65 Layout
-
