--- keyboards/aeboards/ext65/rev2/readme.md
@@ -1,13 +1,0 @@
-# Ext65Rev2
-
-Ext65Rev2 Keyboard
-
-* Keyboard Maintainer: [Xelus22](https://github.com/Xelus22)
-* Hardware Supported: Ext65Rev2
-* Hardware Availability: [Keyboard Treehouse](https://keyboardtreehouse.com/collections/all/products/ext65-r2-pcb)
-
-Make example for this keyboard (after setting up your build environment):
-
-    make aeboards/ext65/rev2:default
-
-See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
