+++ keyboards/aeboards/ext65/rev2/rev2.c
@@ -1,5 +1,8 @@
 #include "rev2.h"
 
+// Tested and verified working on ext65rev2
+void matrix_io_delay(void) { __asm__ volatile("nop\nnop\nnop\n"); }
+
 #ifdef OLED_DRIVER_ENABLE
 void board_init(void) {
   SYSCFG->CFGR1 |= SYSCFG_CFGR1_I2C1_DMA_RMP;
